Server mouse mode: improve user experience
Best server mode mouse experience is given when capturing the mouse,
which makes all the virtual desktop area reachable by the pointer.
Mouse lock has to be triggered by a user action for security reasons, so
we make the user click on the eventCanvas to capture the mouse.

Before, the user could use the mouse without capturing it.
Now, the user always has to click on the canvas when in server mtart using
the mouse when in server mouse mode.